Update footer with select-box interactivity for mobile/tablet resolution
